Fix Debian test flapping
I think the cause of the flapping problem is that we do a check to see if we can get easy_setup
from bitbucket and if we cannot, we get it from GH. For some reason, sometimes bitbucket.org is
not responding. (Maybe we're getting throttled?)

The key here is that these two files are *not* the same. It turns out that the setuptools folks
decided to prepare for the release of setuptools 3.0 by changing the URL to the
unreleased version in their *master* branch version of easy_setup.py.

This change directs any request to GH to use the 2.2 tag of the script. When 3.0 comes out, we
may wish to re-address this.