Remove aggressive checking for Fedora version
Fedora 22 and 23 have reached EOL and we're not testing against
those versions anymore. We can remove the checks for those versions.

We are, however, testing Fedora 25. Many of the checks for installing
Fedora packages were looking specifically for Fedora 23 or 24, but not
leaving room for newer versions. Now that the transition from yum to
dnf has settled down, let's just look for "Fedora" in all of these states
instead of specific versions of Fedora.

This should also fix the installation failures seen against running these
states for Fedora 25.