Always execute both pip install states on Fedora
These states were limited to running based on the output of the
`onlyif` command. But on Fedora, there is a weird bug where the
pip3 files are removed from `/usr/bin/pip*`. We can't rely on the
`onlyif` clauses in this case and should always run these two states
when running the `git.salt` states with `pillar="{py3: true}"

Fixes #311