specify exact python version for kitchen command
This will be used for the verifier to figure out which version of python
was setup to run the testing suite.  So this is needed for arch on py2
tests where the python executable is python3.  And then on any of the
other tests where python is python2 and we are setting up py3 to test.