Use pip to install pyopenssl instead of pkg managers
The versions of pyopenssl that are installed via package managers
can be very out of date. This is particularly true for the Ubuntu
and CentOS distros.

The latest version of moto requires a version of pyopenssl >= `0.14`.
Ubuntu packages `0.13`. This is causing about 100 boto tests to fail.

Pyopenssl is on version 17.x in pip, so by using pip to install instead
of the packaged versions, we can run the boto tests and they pass. We
can also catch issues further up the line with a newer pyopenssl package
in our test suite.

Fixes #526