Update PyOpenSSL pip packages on Ubuntu 16
The Ubuntu 16 images already have pyopenssl installed via pip. It
is an old version (0.15) that does not allow the test daemons to
start.

Upgrading the pyopenssl package to match the other distro tests
fixes the issue and makes the Ubuntu 16 test runs consitent with
the other distros.

Every distro should be running the current latest PyOpenSSL pip
package, which is currently 17.3.0.