Pin cffi version to 1.11.0 when running python 3 tests
When the version of pyopenssl was updated in #529 and we switched
to using pip to install pyopenssl, the cffi pip package is also
pulled in. The latest version of cffi and pyopenssl is causing problems
on the python 3 test runs. Pinning the cffi package to a slightly
older package resolves the issue.